backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.0
python-dotenv==1.0.0
alembic==1.13.0
opencc-python-reimplemented==0.1.7
requests==2.31.0
aiofiles==23.2.1
Pillow==10.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Audio processing dependencies
numpy
librosa
scipy
soundfile
mutagen

# Dolphin ASR
dataoceanai-dolphin